refactor(weather): restructure packages and extract WalkScoreCalculator
