+++ src/main/java/com/kakao/termproject/weather/dto/WeatherDetailInternal.java
@@ -1,6 +1,6 @@
 package com.kakao.termproject.weather.dto;
 
-import com.kakao.termproject.weather.WeatherCondition;
+import com.kakao.termproject.weather.domain.WeatherCondition;
 import java.time.LocalDateTime;
 
 public record WeatherDetailInternal(
